requirements.txt
# Tommy Wu 完全自动化求职申请系统 - 依赖安装

# 核心依赖
selenium>=4.15.0
webdriver-manager>=4.0.1
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0

# 浏览器自动化
playwright>=1.40.0

# 数据处理
jsonschema>=4.20.0

# 日志和监控
colorlog>=6.8.0

# AI/LLM集成 (可选)
openai>=1.3.0
anthropic>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
